requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Data sources
yfinance>=0.2.0
gdeltdoc>=0.1.0
requests>=2.31.0

# ML/Finance
# mlfinlab - Optional: Install from GitHub if needed: pip install git+https://github.com/hudson-and-thames/mlfinlab.git
# Note: Triple barrier labeling works without it (has manual fallback implementation)
optuna>=3.0.0

# NLP/Sentiment
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0

# Reporting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
jupyter>=1.0.0
ipykernel>=6.23.0

